Move CalParams, DarkFileInfo, DarkComp, and constants to lrowaccal header
